Refactor UserInput into its own file
