+++ examples/binding_menu.rs
@@ -8,7 +8,7 @@ use bevy_egui::{
     EguiContext, EguiPlugin,
 };
 use derive_more::Display;
-use leafwing_input_manager::{buttonlike::InputButton, prelude::*};
+use leafwing_input_manager::{prelude::*, user_input::InputButton};
 
 const UI_MARGIN: f32 = 10.0;
