+++ src/clashing_inputs.rs
@@ -1,7 +1,8 @@
 //! Handles clashing inputs into a [`InputMap`](crate::input_map::InputMap) in a configurable fashion.
 
-use crate::buttonlike::{InputButton, InputStreams, Timing, UserInput, VirtualButtonState};
+use crate::buttonlike::{Timing, VirtualButtonState};
 use crate::input_map::InputMap;
+use crate::user_input::{InputButton, InputStreams, UserInput};
 use crate::Actionlike;
 use itertools::Itertools;
 use petitset::PetitSet;
