+++ src/lib.rs
@@ -14,13 +14,14 @@ mod display_impl;
 pub mod errors;
 pub mod input_map;
 mod input_mocking;
-// Re-export this at the root level
+// Re-export this at the root level for convenience
 pub use input_mocking::MockInput;
 pub mod axislike;
 pub mod buttonlike;
 pub mod orientation;
 pub mod plugin;
 pub mod systems;
+pub mod user_input;
 
 // Importing the derive macro
 pub use leafwing_input_manager_macros::Actionlike;
@@ -28,9 +29,9 @@ pub use leafwing_input_manager_macros::Actionlike;
 /// Everything you need to get started
 pub mod prelude {
     pub use crate::action_state::{ActionState, ActionStateDriver};
-    pub use crate::buttonlike::UserInput;
     pub use crate::clashing_inputs::ClashStrategy;
     pub use crate::input_map::InputMap;
+    pub use crate::user_input::UserInput;
 
     pub use crate::plugin::DisableInput;
     pub use crate::plugin::InputManagerPlugin;
