+++ src/user_input.rs
@@ -1,0 +1,418 @@
+//! Helpful abstractions over user inputs of all sorts
+
+use bevy_input::{
+    gamepad::{Gamepad, GamepadButton, GamepadButtonType},
+    keyboard::KeyCode,
+    mouse::MouseButton,
+    Input,
+};
+
+use bevy_utils::HashSet;
+use petitset::PetitSet;
+use serde::{Deserialize, Serialize};
+
+/// Some combination of user input, which may cross [`Input`] boundaries
+///
+/// Suitable for use in an [`InputMap`](crate::input_map::InputMap)
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub enum UserInput {
+    /// A single button
+    Single(InputButton),
+    /// A combination of buttons, pressed simultaneously
+    ///
+    /// Up to 8 (!!) buttons can be chorded together at once.
+    /// Chords are considered to belong to all of the [InputMode]s of their constituent buttons.
+    Chord(PetitSet<InputButton, 8>),
+}
+
+impl UserInput {
+    /// Creates a [`UserInput::Chord`] from an iterator of [`Button`]s
+    ///
+    /// If `buttons` has a length of 1, a [`UserInput::Single`] variant will be returned instead.
+    pub fn chord(buttons: impl IntoIterator<Item = impl Into<InputButton>>) -> Self {
+        // We can't just check the length unless we add an ExactSizeIterator bound :(
+        let mut length: u8 = 0;
+
+        let mut set: PetitSet<InputButton, 8> = PetitSet::default();
+        for button in buttons {
+            length += 1;
+            set.insert(button.into());
+        }
+
+        match length {
+            1 => UserInput::Single(set.into_iter().next().unwrap()),
+            _ => UserInput::Chord(set),
+        }
+    }
+
+    /// Which [`InputMode`]s does this input contain?
+    pub fn input_modes(&self) -> PetitSet<InputMode, 3> {
+        let mut set = PetitSet::default();
+        match self {
+            UserInput::Single(button) => {
+                set.insert((*button).into());
+            }
+            UserInput::Chord(buttons) => {
+                for &button in buttons.iter() {
+                    set.insert(button.into());
+                }
+            }
+        }
+        set
+    }
+
+    /// Does this [`UserInput`] match the provided [`InputMode`]?
+    ///
+    /// For [`UserInput::Chord`], this will be true if any of the buttons in the combination match.
+    pub fn matches_input_mode(&self, input_mode: InputMode) -> bool {
+        // This is slightly faster than using Self::input_modes
+        // As we can return early
+        match self {
+            UserInput::Single(button) => {
+                let button_mode: InputMode = (*button).into();
+                button_mode == input_mode
+            }
+            UserInput::Chord(set) => {
+                for button in set.iter() {
+                    let button_mode: InputMode = (*button).into();
+                    if button_mode == input_mode {
+                        return true;
+                    }
+                }
+                false
+            }
+        }
+    }
+
+    /// The number of buttons in the [`UserInput`]
+    pub fn len(&self) -> usize {
+        match self {
+            UserInput::Single(_) => 1,
+            UserInput::Chord(button_set) => button_set.len(),
+        }
+    }
+
+    /// Is the number of buttons in the [`UserInput`] 0?
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+
+    /// How many of the provided `buttons` are found in the [`UserInput`]
+    ///
+    /// # Example
+    /// ```rust
+    /// use bevy_input::keyboard::KeyCode::*;
+    /// use bevy_utils::HashSet;
+    /// use leafwing_input_manager::buttonlike::UserInput;
+    ///
+    /// let buttons = HashSet::from_iter([LControl.into(), LAlt.into()]);
+    /// let a: UserInput  = A.into();
+    /// let ctrl_a = UserInput::chord([LControl, A]);
+    /// let ctrl_alt_a = UserInput::chord([LControl, LAlt, A]);
+    ///
+    /// assert_eq!(a.n_matching(&buttons), 0);
+    /// assert_eq!(ctrl_a.n_matching(&buttons), 1);
+    /// assert_eq!(ctrl_alt_a.n_matching(&buttons), 2);
+    /// ```
+    pub fn n_matching(&self, buttons: &HashSet<InputButton>) -> usize {
+        match self {
+            UserInput::Single(button) => {
+                if buttons.contains(button) {
+                    1
+                } else {
+                    0
+                }
+            }
+            UserInput::Chord(chord_buttons) => {
+                let mut n_matching = 0;
+                for button in buttons.iter() {
+                    if chord_buttons.contains(button) {
+                        n_matching += 1;
+                    }
+                }
+
+                n_matching
+            }
+        }
+    }
+}
+
+impl From<InputButton> for UserInput {
+    fn from(input: InputButton) -> Self {
+        UserInput::Single(input)
+    }
+}
+
+impl From<GamepadButtonType> for UserInput {
+    fn from(input: GamepadButtonType) -> Self {
+        UserInput::Single(InputButton::Gamepad(input))
+    }
+}
+
+impl From<KeyCode> for UserInput {
+    fn from(input: KeyCode) -> Self {
+        UserInput::Single(InputButton::Keyboard(input))
+    }
+}
+
+impl From<MouseButton> for UserInput {
+    fn from(input: MouseButton) -> Self {
+        UserInput::Single(InputButton::Mouse(input))
+    }
+}
+
+/// A button-like input type
+///
+/// See [`Button`] for the value-ful equivalent.
+/// Use the [`From`] or [`Into`] traits to convert from a [`InputButton`] to a [`InputMode`].
+///
+/// Unfortunately we cannot use a trait object here, as the types used by `Input`
+/// require traits that are not object-safe.
+///
+/// Please contact the maintainers if you need support for another type!
+#[non_exhaustive]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum InputMode {
+    /// A gamepad
+    Gamepad,
+    /// A keyboard
+    Keyboard,
+    /// A mouse
+    Mouse,
+}
+
+impl InputMode {
+    /// Iterates over the possible [`InputModes`](InputMode)
+    pub fn iter() -> InputModeIter {
+        InputModeIter::default()
+    }
+}
+
+/// An iterator of [`InputModes`](InputMode)
+///
+/// Created by calling [`InputMode::iter`]
+#[derive(Debug, Clone, Default)]
+pub struct InputModeIter {
+    cursor: u8,
+}
+
+impl Iterator for InputModeIter {
+    type Item = InputMode;
+
+    fn next(&mut self) -> Option<InputMode> {
+        let item = match self.cursor {
+            0 => Some(InputMode::Gamepad),
+            1 => Some(InputMode::Keyboard),
+            2 => Some(InputMode::Mouse),
+            _ => None,
+        };
+        if self.cursor <= 2 {
+            self.cursor += 1;
+        }
+
+        item
+    }
+}
+
+impl From<InputButton> for InputMode {
+    fn from(button: InputButton) -> Self {
+        match button {
+            InputButton::Gamepad(_) => InputMode::Gamepad,
+            InputButton::Keyboard(_) => InputMode::Keyboard,
+            InputButton::Mouse(_) => InputMode::Mouse,
+        }
+    }
+}
+
+/// The values of a button-like input type
+///
+/// See [`InputMode`] for the value-less equivalent. Commonly stored in the [`UserInput`] enum.
+///
+/// Unfortunately we cannot use a trait object here, as the types used by `Input`
+/// require traits that are not object-safe.
+///
+/// Please contact the maintainers if you need support for another type!
+#[non_exhaustive]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub enum InputButton {
+    /// A button on a gamepad
+    Gamepad(GamepadButtonType),
+    /// A button on a keyboard
+    Keyboard(KeyCode),
+    /// A button on a mouse
+    Mouse(MouseButton),
+}
+
+impl From<GamepadButtonType> for InputButton {
+    fn from(input: GamepadButtonType) -> Self {
+        InputButton::Gamepad(input)
+    }
+}
+
+impl From<KeyCode> for InputButton {
+    fn from(input: KeyCode) -> Self {
+        InputButton::Keyboard(input)
+    }
+}
+
+impl From<MouseButton> for InputButton {
+    fn from(input: MouseButton) -> Self {
+        InputButton::Mouse(input)
+    }
+}
+
+/// A collection of [`Input`] structs, which can be used to update an [`InputMap`](crate::input_map::InputMap).
+///
+/// Each of these streams is optional; if a stream does not exist, it is treated as if it were entirely unpressed.
+///
+/// These are typically collected via a system from the [`World`](bevy::prelude::World) as resources.
+#[derive(Debug, Clone)]
+pub struct InputStreams<'a> {
+    /// An optional [`GamepadButton`] [`Input`] stream
+    pub gamepad: Option<&'a Input<GamepadButton>>,
+    /// An optional [`KeyCode`] [`Input`] stream
+    pub keyboard: Option<&'a Input<KeyCode>>,
+    /// An optional [`MouseButton`] [`Input`] stream
+    pub mouse: Option<&'a Input<MouseButton>>,
+    /// The [`Gamepad`] that this struct will detect inputs from
+    pub associated_gamepad: Option<Gamepad>,
+}
+
+// Constructors
+impl<'a> InputStreams<'a> {
+    /// Construct [`InputStreams`] with only a [`GamepadButton`] input stream
+    pub fn from_gamepad(
+        gamepad_input_stream: &'a Input<GamepadButton>,
+        associated_gamepad: Gamepad,
+    ) -> Self {
+        Self {
+            gamepad: Some(gamepad_input_stream),
+            keyboard: None,
+            mouse: None,
+            associated_gamepad: Some(associated_gamepad),
+        }
+    }
+
+    /// Construct [`InputStreams`] with only a [`KeyCode`] input stream
+    pub fn from_keyboard(keyboard_input_stream: &'a Input<KeyCode>) -> Self {
+        Self {
+            gamepad: None,
+            keyboard: Some(keyboard_input_stream),
+            mouse: None,
+            associated_gamepad: None,
+        }
+    }
+
+    /// Construct [`InputStreams`] with only a [`GamepadButton`] input stream
+    pub fn from_mouse(mouse_input_stream: &'a Input<MouseButton>) -> Self {
+        Self {
+            gamepad: None,
+            keyboard: None,
+            mouse: Some(mouse_input_stream),
+            associated_gamepad: None,
+        }
+    }
+}
+
+// Input checking
+impl<'a> InputStreams<'a> {
+    /// Is the `input` matched by the [`InputStreams`]?
+    pub fn input_pressed(&self, input: &UserInput) -> bool {
+        match input {
+            UserInput::Single(button) => self.button_pressed(*button),
+            UserInput::Chord(buttons) => self.all_buttons_pressed(buttons),
+        }
+    }
+
+    /// Is at least one of the `inputs` pressed?
+    #[must_use]
+    pub fn any_pressed(&self, inputs: &PetitSet<UserInput, 16>) -> bool {
+        for input in inputs.iter() {
+            if self.input_pressed(input) {
+                return true;
+            }
+        }
+        // If none of the inputs matched, return false
+        false
+    }
+
+    /// Is the `button` pressed?
+    #[must_use]
+    pub fn button_pressed(&self, button: InputButton) -> bool {
+        match button {
+            InputButton::Gamepad(gamepad_button) => {
+                // If no gamepad is registered, we know for sure that no match was found
+                if let Some(gamepad) = self.associated_gamepad {
+                    if let Some(gamepad_stream) = self.gamepad {
+                        gamepad_stream.pressed(GamepadButton(gamepad, gamepad_button))
+                    } else {
+                        false
+                    }
+                } else {
+                    false
+                }
+            }
+            InputButton::Keyboard(keycode) => {
+                if let Some(keyboard_stream) = self.keyboard {
+                    keyboard_stream.pressed(keycode)
+                } else {
+                    false
+                }
+            }
+            InputButton::Mouse(mouse_button) => {
+                if let Some(mouse_stream) = self.mouse {
+                    mouse_stream.pressed(mouse_button)
+                } else {
+                    false
+                }
+            }
+        }
+    }
+
+    /// Are all of the `buttons` pressed?
+    #[must_use]
+    pub fn all_buttons_pressed(&self, buttons: &PetitSet<InputButton, 8>) -> bool {
+        for &button in buttons.iter() {
+            // If any of the appropriate inputs failed to match, the action is considered pressed
+            if !self.button_pressed(button) {
+                return false;
+            }
+        }
+        // If none of the inputs failed to match, return true
+        true
+    }
+}
+
+/// A mutable collection of [`Input`] structs, which can be used for mocking user inputs.
+///
+/// Each of these streams is optional; if a stream does not exist, inputs sent to them will be ignored.
+///
+/// These are typically collected via a system from the [`World`](bevy::prelude::World) as resources.
+#[derive(Debug)]
+pub struct MutableInputStreams<'a> {
+    /// An optional [`GamepadButton`] [`Input`] stream
+    pub gamepad: Option<&'a mut Input<GamepadButton>>,
+    /// An optional [`KeyCode`] [`Input`] stream
+    pub keyboard: Option<&'a mut Input<KeyCode>>,
+    /// An optional [`MouseButton`] [`Input`] stream
+    pub mouse: Option<&'a mut Input<MouseButton>>,
+    /// The [`Gamepad`] that this struct will detect inputs from
+    pub associated_gamepad: Option<Gamepad>,
+}
+
+impl<'a> From<MutableInputStreams<'a>> for InputStreams<'a> {
+    fn from(mutable_streams: MutableInputStreams<'a>) -> Self {
+        let gamepad = mutable_streams.gamepad.map(|mutable_ref| &*mutable_ref);
+
+        let keyboard = mutable_streams.keyboard.map(|mutable_ref| &*mutable_ref);
+
+        let mouse = mutable_streams.mouse.map(|mutable_ref| &*mutable_ref);
+
+        InputStreams {
+            gamepad,
+            keyboard,
+            mouse,
+            associated_gamepad: mutable_streams.associated_gamepad,
+        }
+    }
+}
